+++ packages/client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @airstate/client
 
+## 0.1.12
+
+### Patch Changes
+
+- 241cba5: Improved presence type exported. Localized to @airstate/client and not re-exported from @airstate/server
+
+## 0.1.11
+
+### Patch Changes
+
+- 52fa7f2: added READMEs
+
 ## 0.1.10
 
 ### Patch Changes